Added pubport and stackpointer and cleanup
pubport refferes to the public port of the endpoint, while the
stackpointer reffers to the where on the stack the tree refers.

Stackpointer is not needed, but if a relation is made, it must be
configured if not set in endpoint. If it is configured during relation
of tree, it is not needed here.